api/requirements.txt
tensorflow==2.18.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.18.0
matplotlib
numpy==1.26.4